core/requirements.txt
Flask~=3.1.2
kafka-python>=2.0.2
pymongo>=4.0